# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Communication_LTD Ticket System API
  description: |
    API for Communication_LTD customer support ticket system
    
    ## Logging Requirements
    
    - All successful mutating operations (POST, PUT, PATCH, DELETE) with valid CSRF tokens are logged at INFO level
    - Missing or invalid CSRF tokens are logged at WARNING level
    - Requests approaching rate limit thresholds are logged at WARNING level
    - All requests include client IP, user ID (if authenticated), timestamp, and operation details
    
    ## Rate Limiting
    
    The API implements rate limiting to protect against abuse:
    - Authentication endpoints: 10 requests per minute per IP
    - Standard endpoints: 60 requests per minute per authenticated user
    - Admin endpoints: 30 requests per minute per admin user
    
    When requests approach 80% of their limit threshold, WARNING level logs are generated
  version: 1.0.0
  contact:
    name: Communication_LTD Support
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: >
        CSRF token for mutation requests protection.
        All POST, PUT, PATCH, DELETE operations require this header.
        Missing or invalid tokens will be logged at WARNING level.
        Valid tokens on successful mutating requests will be logged at INFO level.
      required: true
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      required:
        - code
        - message
      example:
        code: 400
        message: Bad Request
        details:
          field: "username"
          reason: "Username is already taken"
...
    Package:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        features:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - created_at
        - updated_at
      example:
        id: 1
        name: "Basic Internet"
        description: "Basic internet package for home use"
        price: 29.99
        features:
          - "50Mbps download speed"
          - "10Mbps upload speed"
          - "Unlimited data"
        created_at: "2025-08-01T10:00:00Z"
        updated_at: "2025-08-01T10:00:00Z"
    PackageCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        features:
          type: array
          items:
            type: string
      required:
        - name
        - price
      example:
        name: "Basic Internet"
        description: "Basic internet package for home use"
        price: 29.99
        features:
          - "50Mbps download speed"
          - "10Mbps upload speed"
          - "Unlimited data"
    CustomerPackage:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customer_id:
          type: integer
          format: int64
        package_id:
          type: integer
          format: int64
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [active, expired, cancelled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - customer_id
        - package_id
        - start_date
        - status
        - created_at
        - updated_at
      example:
        id: 1
        customer_id: 1
        package_id: 1
        start_date: "2025-08-01"
        end_date: "2026-08-01"
        status: "active"
        created_at: "2025-08-01T10:00:00Z"
        updated_at: "2025-08-01T10:00:00Z"
    CustomerPackageCreate:
      type: object
      properties:
        package_id:
          type: integer
          format: int64
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [active, expired, cancelled]
          default: active
      required:
        - package_id
        - start_date
      example:
        package_id: 1
        start_date: "2025-08-01"
        end_date: "2026-08-01"
        status: "active"
...
paths:
...
  /customers/{customer_id}/packages:
    parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get customer's packages
      operationId: getCustomerPackages
      tags:
        - Customers
        - Packages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of customer packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerPackage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Add a package to customer
      operationId: addCustomerPackage
      tags:
        - Customers
        - Packages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerPackageCreate'
      responses:
        '201':
          description: Package added to customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPackage'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Customer or package not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /packages:
    get:
      summary: List all packages
      operationId: listPackages
      tags:
        - Packages
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: List of packages
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Package'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new package
      operationId: createPackage
      tags:
        - Packages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageCreate'
      responses:
        '201':
          description: Package created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /packages/{package_id}:
    parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get package details
      operationId: getPackage
      tags:
        - Packages
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Package details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Package not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update package details
      operationId: updatePackage
      tags:
        - Packages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageCreate'
      responses:
        '200':
          description: Package updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Package'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Package not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a package
      operationId: deletePackage
      tags:
        - Packages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Package deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Package not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...